refactor: migrate LIBERO RPC adapters to base classes
