+++ docs/source-en/rst_source/development/interfaces.rst
@@ -7,7 +7,8 @@ Walkthroughs: :doc:`add_robot`, :doc:`add_primitive`. Repo layout: :doc:`archite
 Environment entry
 -----------------
 
-After you add ``robots/<env>/``, ``main.py`` calls two functions in ``__init__.py``:
+After you add ``robots/<env>/``, the package ``__init__.py`` re-exports two
+functions implemented in ``env_spec.py`` for ``main.py`` to call:
 
 .. code-block:: python
 
@@ -51,7 +52,7 @@ After you add ``robots/<env>/``, ``main.py`` calls two functions in ``__init__.p
 ``dashboard_events`` and ``video_path`` are supplied by the active runner, so
 you normally do not need to change them.
 
-References: ``robots/libero/__init__.py`` and ``robots/libero/spec.py``.
+Reference: ``robots/libero/env_spec.py``.
 
 Planner
 -------
@@ -125,8 +126,11 @@ for large or history-stacked nested-NumPy observations to move length-prefixed
 pickle frames and skip repeated JSON encoding. Pickle is unsafe on untrusted
 input, so only point ``socket`` at trusted endpoints.
 
-Server: subclass ``rpent.utils.rpc.RpcFacade`` and implement ``_dispatch`` for
-business RPCs (e.g. ``reset``, ``step``, ``predict``). Do not implement ``healthz`` or
-``shutdown`` in the subclass.
+Environment and VLA clients should normally subclass ``BaseEnvClient`` and
+``BaseVLAClient``; their servers should subclass ``BaseEnvFacade`` and
+``BaseVLAFacade`` and register extension routes through ``_register_rpc``. The
+bases provide common routing and locking on top of ``RpcFacade``. Subclass
+``RpcFacade`` directly only for a service type without a specialized base. Do
+not implement ``healthz`` or ``shutdown`` in application subclasses.
 
 Details are in the env_server / vla_server sections of :doc:`add_robot`.
