+++ docs/source-zh/rst_source/development/interfaces.rst
@@ -7,7 +7,8 @@
 环境入口
 --------
 
-把包放到 ``robots/<env>/`` 后，``main.py`` 会调用 ``__init__.py`` 里的两个函数：
+把包放到 ``robots/<env>/`` 后，包的 ``__init__.py`` 会重导出
+``env_spec.py`` 中实现的两个函数，供 ``main.py`` 调用：
 
 .. code-block:: python
 
@@ -49,7 +50,7 @@
 ``get_toolkit`` 一般只需把 ``primitives_kwargs`` 传给环境子类；
 ``dashboard_events``、``video_path`` 由当前 runner 传入，通常不用改。
 
-参考实现：``robots/libero/__init__.py`` 和 ``robots/libero/spec.py``。
+参考实现：``robots/libero/env_spec.py``。
 
 Planner
 -------
@@ -117,7 +118,9 @@ runtime 钩子中解析）：
 观测数据很大、或是多帧堆叠的嵌套 NumPy 字典时可改 ``socket``，用带长度前缀的
 pickle 数据帧传输，省掉反复的 JSON 编解码。pickle 不适合不可信输入，socket 只应连接可信端点。
 
-服务端：继承 ``rpent.utils.rpc.RpcFacade``，实现 ``_dispatch`` 分发业务 RPC
-（如 ``reset``、``step``、``predict``）。``healthz`` / ``shutdown`` 不必在子类里写。
+环境和 VLA client 通常应分别继承 ``BaseEnvClient``、``BaseVLAClient``；服务端
+分别继承 ``BaseEnvFacade``、``BaseVLAFacade``，并通过 ``_register_rpc`` 注册
+扩展路由。这些基类在 ``RpcFacade`` 之上提供公共路由和锁。只有尚无专用基类的
+服务类型才直接继承 ``RpcFacade``。业务子类不必实现 ``healthz`` / ``shutdown``。
 
 细节见 :doc:`add_robot` 中的 env_server 与 vla_server 章节。
